feat: add error handlers in update and delete methods in product service

+++ src/main/java/tech/joelf/e_commerce_api/services/ProductService.java
@@ -2,12 +2,15 @@ package tech.joelf.e_commerce_api.services;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.BeanUtils;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import jakarta.persistence.EntityNotFoundException;
 import tech.joelf.e_commerce_api.dtos.request.ProductDtoIn;
 import tech.joelf.e_commerce_api.dtos.response.ProductDtoOut;
+import tech.joelf.e_commerce_api.exceptions.DatabaseException;
 import tech.joelf.e_commerce_api.exceptions.ResourceNotFoundException;
 import tech.joelf.e_commerce_api.models.Product;
 import tech.joelf.e_commerce_api.repositories.ProductRepository;
@@ -40,13 +43,25 @@ public class ProductService {
     }
 
     public ProductDtoOut update(Long id, ProductDtoIn dto) {
-        var product = productRepository.getReferenceById(id);
-        BeanUtils.copyProperties(modelMapper.map(dto, Product.class), product, "id");
+        try {
+            var product = productRepository.getReferenceById(id);
+            BeanUtils.copyProperties(modelMapper.map(dto, Product.class), product, "id");
 
-        return modelMapper.map(productRepository.save(product), ProductDtoOut.class);
+            return modelMapper.map(productRepository.save(product), ProductDtoOut.class);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Product not found.");
+        }
     }
 
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Product not found.");
+        }
+
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Referencial integrity violated.");
+        }
     }
 }
